+++ src/Entity/Feedback.php
@@ -1,0 +1,134 @@
+<?php
+
+namespace App\Entity;
+
+use App\Repository\FeedbackRepository;
+use Doctrine\DBAL\Types\Types;
+use Doctrine\ORM\Mapping as ORM;
+use App\Contract\SerializableInterface;
+
+#[ORM\HasLifecycleCallbacks]
+#[ORM\Entity(repositoryClass: FeedbackRepository::class)]
+class Feedback implements SerializableInterface
+{
+
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column]
+    private ?int $id = null;
+
+    #[ORM\Column(length: 255)]
+    private ?string $rating = null;
+
+    #[ORM\Column(type: Types::TEXT, nullable: true)]
+    private ?string $comment = null;
+
+    #[ORM\Column(length: 255)]
+    private ?string $context = null;
+
+    #[ORM\Column(length: 255)]
+    private ?string $contextId = null;
+
+    #[ORM\Column]
+    private ?\DateTimeImmutable $createdAt = null;
+
+    #[ORM\ManyToOne]
+    #[ORM\JoinColumn(nullable: false, onDelete: 'CASCADE')]
+    private ?User $author = null;
+
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    public function getRating(): ?string
+    {
+        return $this->rating;
+    }
+
+    public function setRating(string $rating): static
+    {
+        $this->rating = $rating;
+
+        return $this;
+    }
+
+    public function getComment(): ?string
+    {
+        return $this->comment;
+    }
+
+    public function setComment(?string $comment): static
+    {
+        $this->comment = $comment;
+
+        return $this;
+    }
+
+    public function getContext(): ?string
+    {
+        return $this->context;
+    }
+
+    public function setContext(string $context): static
+    {
+        $this->context = $context;
+
+        return $this;
+    }
+
+    public function getContextId(): ?string
+    {
+        return $this->contextId;
+    }
+
+    public function setContextId(string $contextId): static
+    {
+        $this->contextId = $contextId;
+
+        return $this;
+    }
+
+    public function getCreatedAt(): ?\DateTimeImmutable
+    {
+        return $this->createdAt;
+    }
+
+    public function setCreatedAt(\DateTimeImmutable $createdAt): static
+    {
+        $this->createdAt = $createdAt;
+
+        return $this;
+    }
+
+    public function getAuthor(): ?User
+    {
+        return $this->author;
+    }
+
+    public function setAuthor(?User $author): static
+    {
+        $this->author = $author;
+
+        return $this;
+    }
+
+    #[ORM\PrePersist]
+    public function setCreatedAtValue(): void
+    {
+        $this->createdAt = new \DateTimeImmutable();
+    }
+
+    public function serialize(): array
+    {
+        return [
+            'id' => $this->id,
+            'rating' => $this->rating,
+            'context' => $this->context,
+            'contextId' => $this->contextId,
+            'createdAt' => $this->createdAt->format('Y-m-d H:i:s'),
+            'author' => $this->author->serialize(),
+            'comment' => $this->comment,
+        ];
+    }
+}